+++ firmware/statemachine/color.h
@@ -1,0 +1,36 @@
+// The MIT License (MIT)
+
+// Copyright (c) 2016 Reid Levenick
+
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+#ifndef STATEMACHINE_COLOR_H
+#define STATEMACHINE_COLOR_H
+
+#include <stdint.h>
+
+#include "constants.h"
+
+typedef struct Color_t {
+    uint8_t r; uint8_t g; uint8_t b;
+} Color;
+
+void Color_emplace(Color*, uint8_t, uint8_t, uint8_t);
+void Color_interpolate(Color*, Color, Color, Interpolation, uint8_t);
+
+#endif
